test_push_notifications: Split HandlePushNotificationTest to a new file.
This commit moves `HandlePushNotificationTest` to a new
`test_handle_push_notification.py` file.

It helps in making it easier to work with `test_push_notifications.py`
which was 4000+ lines of code.
